requirements.txt
pygsheets
pandas
python-dotenv
numpy
